LPS-83992 Merge MethodSignature into MethodType, so we don't have to manually ensure using them in the right pair.
